Move middleware to middleware.py
