+++ python/nav/django/settings.py
@@ -127,8 +127,8 @@ TEMPLATES = [
 MIDDLEWARE = (
     'django.middleware.common.CommonMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
-    'nav.web.auth.AuthenticationMiddleware',
-    'nav.web.auth.AuthorizationMiddleware',
+    'nav.web.auth.middleware.AuthenticationMiddleware',
+    'nav.web.auth.middleware.AuthorizationMiddleware',
     'nav.django.legacy.LegacyCleanupMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
 )
